Added a resize function to the video handler. Added documentation comments to the video handler file. #60 #95

+++ TeachEasy/src/teacheasy/render/Renderer.java
@@ -168,8 +168,8 @@ public class Renderer {
     
     /** Render a video on a page */
     private void renderVideo(VideoObject video) {
-        videoHandler.createVideo(bounds.getMaxX() * video.getXStart(),
-                                 bounds.getMaxY() * video.getYStart(),
+        videoHandler.createVideo((float)bounds.getMaxX() * video.getXStart(),
+                                 (float)bounds.getMaxY() * video.getYStart(),
                                  800,
                                  video.getSourcefile(),
                                  false,
